lectures/06-github-pipelines/Testing codes/requirements.txt
stages:
  - test

test:
  stage: test
  script:
    - pip install -r requirements.txt
    - pytest test_main_code.py
